requirements.txt
Django
Pillow
jsonschema
django-ratelimit
django-widget-tweaks